Use `int8` instead of `integer` for permssions table

+++ crates/adapter/repositories_sqlx/migrations/20221110023431_create_permissions_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE permissions
     id UUID NOT NULL DEFAULT gen_random_uuid() PRIMARY KEY,
 
     resource BIGINT NOT NULL,
-    actions INTEGER NOT NULL,
+    actions INT8 NOT NULL,
     role_id UUID NOT NULL,
     created_at TIMESTAMPTZ DEFAULT NOW() NOT NULL,
